requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.40.0
python-multipart==0.0.12
pypdf==4.3.0
python-docx==1.1.2
